Long Night. Added terminal Velocity. Increasing speed by gravity's constant value minused by thrust.
Simplified the Switching between when there is no fuel left.
To let gravity take over.

Added var for crash speed to change how fast was "TooFast" on touch down.

Added bounce mechanism, When lander hits the top of the screen the Trackbar slider jumps down the slider, as if the control got bumped.

messed around with the dividing num of the Trackbar.value... Decided on 15.